\name{networkSpatial-package}
\alias{networkSpatial-package}
\alias{networkSpatial}
\docType{package}
\title{Tools for the Generation and Analysis of Spatially-embedded Networks}
\description{
Provides tools for simulation and analysis of spatially-embedded networks, including spatial Bernoulli graph generation, SIF inference, and more.
}
\details{
\tabular{ll}{
Package: \tab networkSpatial\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2011-03-25\cr
License: \tab GPL (>=2) \cr
LazyLoad: \tab yes\cr
}

}
\author{
Carter T. Butts <buttsc@uci.edu>, Zack W. Almquist <almquist@umn.edu>
Maintainer: Carter T. Butts <buttsc@uci.edu>

}
\references{
Butts, Carter T.  (forthcoming).  Space and Structure: Models and Methods for Large-Scale Interpersonal Networks.  Springer.  
}

\keyword{ package }
%\seealso{}
%\examples{}
